Expand parameterization 3 sigma values to 64 bits width
cgbn: widen the stage-1 sigma argument to uint64_t

Change cgbn_ecm_stage1's starting sigma from uint32_t to uint64_t so a
param-3 batch can retain all 64 numerator bits across the C/CUDA boundary.
The curve count remains uint32_t; the denominator remains 2^32.

This changes the internal function interface. Apply it together with
the definition in cgbn_stage1.cu and the caller in cudawrapper.c, and
rebuild the affected objects. The ecm_params layout and save syntax
are unchanged.

+++ cgbn_stage1.h
@@ -31,7 +31,7 @@ extern "C" {
 
 int cgbn_ecm_stage1(mpz_t *factors, int *array_found,
              const mpz_t N, const mpz_t s,
-             uint32_t curves, uint32_t sigma,
+             uint32_t curves, uint64_t sigma,
              float *gputime, int verbose);
 
 #ifdef __cplusplus
